+++ src/exchanges/gateio_usd_futures/exchange.rs
@@ -161,7 +161,7 @@ impl Exchange for GateioUSDFuturesExchange {
         ));
     }
 
-    fn cancel_order(&self, _order_id: String) -> () {
+    fn cancel_order(&self, _order: Order) -> () {
         let _ = self.logs_sender.send(Log::new(
             LogLevel::Warning("NA".to_string(), None),
             format!("Trading is not supported on {}", self.name()).to_string(),